+++ src/main/java/codes/thischwa/c5c/requestcycle/response/mode/SaveFile.java
@@ -27,7 +27,7 @@ public final class SaveFile extends GenericResponse {
 		this.fullPath = fullPath;
 	}
 	
-	@JsonProperty("Path")
+	@JsonProperty("Builder")
 	public String getFullPath() {
 		return fullPath;
 	}