+++ src/test/java/codes/thischwa/c5c/requestcycle/response/mode/FolderInfoTest.java
@@ -24,15 +24,15 @@ public class FolderInfoTest {
 	public void testToString() {
 		FolderInfo resp = new FolderInfo();
 		
-		FileInfo fileInfo = new FileInfo("/tmp/", false);
-		fileInfo.setFileProperties(new FileInfoProperties("img01.png", 200, 100, 30024, null));
+		FileInfo fileInfo = new FileInfo("/tmp/", true, false);
+		fileInfo.setFileProperties(new FileInfoProperties("img01.png", false, 200, 100, 30024, null));
 		resp.add(fileInfo);
 		
-		fileInfo = new FileInfo("/tmp", true);
-		fileInfo.setFileProperties(new FileInfoProperties("folder", null));
+		fileInfo = new FileInfo("/tmp", true, false);
+		fileInfo.setFileProperties(new FileInfoProperties("folder", false, null));
 		resp.add(fileInfo);
 
-		String expected = "{\"\\/tmp\\/img01.png\":{\"Capabilities\":[],\"Code\":0,\"Error\":\"\",\"File Type\":\"png\",\"Filename\":\"img01.png\",\"Path\":\"\\/tmp\\/img01.png\",\"Preview\":null,\"Properties\":{\"Date Created\":null,\"Date Modified\":null,\"Height\":100,\"Size\":30024,\"Width\":200}},\"\\/tmp\\/folder\\/\":{\"Capabilities\":[],\"Code\":0,\"Error\":\"\",\"File Type\":\"dir\",\"Filename\":\"folder\",\"Path\":\"\\/tmp\\/folder\\/\",\"Preview\":null,\"Properties\":{\"Date Created\":null,\"Date Modified\":null,\"Height\":null,\"Size\":null,\"Width\":null}}}";
+		String expected = "{\"\\/tmp\\/img01.png\":{\"Capabilities\":[],\"Code\":0,\"Error\":\"\",\"File Type\":\"png\",\"Filename\":\"img01.png\",\"PathBuilder\":\"\\/tmp\\/img01.png\",\"Preview\":null,\"Properties\":{\"Date Created\":null,\"Date Modified\":null,\"Height\":100,\"Size\":30024,\"Width\":200},\"Protected\":0},\"\\/tmp\\/folder\\/\":{\"Capabilities\":[],\"Code\":0,\"Error\":\"\",\"File Type\":\"dir\",\"Filename\":\"folder\",\"PathBuilder\":\"\\/tmp\\/folder\\/\",\"Preview\":null,\"Properties\":{\"Date Created\":null,\"Date Modified\":null,\"Height\":null,\"Size\":null,\"Width\":null},\"Protected\":0}}";
 		assertEquals(expected, resp.toString());
 	}